Refactor: decouple CashflowSankey from MonthlySpending via shared query (fixes #49)
