+++ tests/features/CashflowSankey/service.test.ts
@@ -1,38 +1,40 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { CashflowSankeyService } from '../../../backend/src/features/CashflowSankey/service';
-import { MonthlySpendingService } from '../../../backend/src/features/MonthlySpending/service';
+import type { IDbClient } from '../../../backend/src/infrastructure/db/IDbClient';
 
-vi.mock('../../../backend/src/features/MonthlySpending/service', () => {
-  return {
-    MonthlySpendingService: class {
-      getMonthlySpending = vi.fn();
-    }
-  };
-});
+// Mock the shared getMonthlySpending query
+vi.mock('../../../backend/src/shared/queries/getMonthlySpending.js', () => ({
+  getMonthlySpending: vi.fn()
+}));
+
+import { getMonthlySpending } from '../../../backend/src/shared/queries/getMonthlySpending.js';
 
 describe('CashflowSankeyService', () => {
+  let mockDbClient: IDbClient;
   let service: CashflowSankeyService;
 
   beforeEach(() => {
     vi.clearAllMocks();
-    service = new CashflowSankeyService();
+    mockDbClient = { query: vi.fn() };
+    service = new CashflowSankeyService(mockDbClient);
   });
 
   describe('getSankeyData', () => {
     it('should generate nodes and links correctly', async () => {
-      const mockGetMonthlySpending = vi.fn().mockResolvedValue({
+      const queryMock = vi.mocked(getMonthlySpending);
+      queryMock.mockResolvedValue({
         income: [{ Category_Name: 'Salary', amount: 5000 }],
         expenses: [{ Category_Name: 'Rent', amount: 2000 }, { Category_Name: 'Food', amount: 1000 }],
         totalIncome: 5000,
         totalSpent: 3000,
         netIncome: 2000,
-        savingsRate: 40
+        savingsRate: 40,
+        transactions: []
       });
-      (service as any).monthlyService.getMonthlySpending = mockGetMonthlySpending;
 
       const result = await service.getSankeyData('2024-06');
 
-      expect(mockGetMonthlySpending).toHaveBeenCalledWith('2024-06');
+      expect(queryMock).toHaveBeenCalledWith(mockDbClient, '2024-06');
       
       expect(result.nodes.length).toBeGreaterThan(0);
       expect(result.links.length).toBeGreaterThan(0);
@@ -53,15 +55,16 @@ describe('CashflowSankeyService', () => {
     });
 
     it('should include Deficit node when expenses > income', async () => {
-      const mockGetMonthlySpending = vi.fn().mockResolvedValue({
+      const queryMock = vi.mocked(getMonthlySpending);
+      queryMock.mockResolvedValue({
         income: [{ Category_Name: 'Salary', amount: 2000 }],
         expenses: [{ Category_Name: 'Rent', amount: 3000 }],
         totalIncome: 2000,
         totalSpent: 3000,
         netIncome: -1000,
-        savingsRate: 0
+        savingsRate: 0,
+        transactions: []
       });
-      (service as any).monthlyService.getMonthlySpending = mockGetMonthlySpending;
 
       const result = await service.getSankeyData('2024-06');
       const labels = result.nodes.map((n: any) => n.label);
@@ -73,15 +76,16 @@ describe('CashflowSankeyService', () => {
     });
 
     it('should return empty payload if no active income or expenses', async () => {
-      const mockGetMonthlySpending = vi.fn().mockResolvedValue({
+      const queryMock = vi.mocked(getMonthlySpending);
+      queryMock.mockResolvedValue({
         income: [],
         expenses: [],
         totalIncome: 0,
         totalSpent: 0,
         netIncome: 0,
-        savingsRate: 0
+        savingsRate: 0,
+        transactions: []
       });
-      (service as any).monthlyService.getMonthlySpending = mockGetMonthlySpending;
 
       const result = await service.getSankeyData('2024-06');
